task(SDK-5233): fixes pr comments
- uses default long in get long method
- removes /1000 which was left over by mistake.

+++ clevertap-core/src/main/java/com/clevertap/android/sdk/StorageHelper.kt
@@ -118,7 +118,7 @@ internal object StorageHelper {
         context: Context,
         accountId: String,
         rawKey: String,
-        defaultValue: Int,
+        defaultValue: Long,
         nameSpace: String?
     ): Long {
         val key = storageKeyWithSuffix(accountId, rawKey)
